+++ src/openforms/forms/api/serializers/form.py
@@ -305,6 +305,7 @@ class FormSerializer(PublicFieldsSerializerMixin, serializers.ModelSerializer):
             "submission_report_download_link_title",
             "brp_personen_request_options",
             "new_renderer_enabled",
+            "new_logic_evaluation_enabled",
         )
         # allowlist for anonymous users
         public_fields = (
@@ -340,6 +341,7 @@ class FormSerializer(PublicFieldsSerializerMixin, serializers.ModelSerializer):
             "submission_statements_configuration",
             "submission_report_download_link_title",
             "new_renderer_enabled",
+            "new_logic_evaluation_enabled",
         )
         extra_kwargs = {
             "uuid": {