+++ src/openforms/forms/models/form.py
@@ -397,6 +397,16 @@ class Form(models.Model):
         default=False,
         help_text=_("Enable the new renderer."),
     )
+    new_logic_evaluation_enabled = models.BooleanField(
+        _("enable new logic rule evaluation"),
+        default=False,
+        help_text=_(
+            "Enabling this will analyze logic rules and re-order them according to "
+            "their dependency on other logic rules (happens when the form is saved). "
+            "Each rule will be automatically assigned to one or more steps on which "
+            "it will be executed."
+        ),
+    )
 
     objects: ClassVar[  # pyright: ignore[reportIncompatibleVariableOverride]
         FormManager